+++ src/core/index.ts
@@ -3,3 +3,4 @@ export * from "./scale";
 export * from "./categories";
 export * from "./convert";
 export * from "./describe";
+export * from "./snap";